:sparkles: Feat: UserInfoComponents.kt 분리
<footer>
- 관련: #13
